requirements.txt
boto3==1.26.41
botocore==1.29.41
jmespath==1.0.1
pydantic==1.10.4
PyJWT==2.6.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.13
